+++ app/models/randomization_algorithm/algorithms/permuted_block.rb
@@ -4,55 +4,11 @@ module RandomizationAlgorithm
   module Algorithms
     # Handles list creation and randomization for permuted block randomization
     # schemes.
-    class PermutedBlock < Default
+    class PermutedBlock < StratifiedLists
       attr_reader :randomization_scheme
 
-      def initialize(randomization_scheme)
-        @randomization_scheme = randomization_scheme
-      end
-
       def randomization_error_message
-        "Block Size Multipliers and Treatment Arms may not be set up correctly."
-      end
-
-      def add_missing_lists!(current_user)
-        stratifications = []
-
-        if number_of_lists.in?(1..RandomizationScheme::MAX_LISTS - 1)
-          stratifications = [[]]
-          @randomization_scheme.stratification_factors.each do |stratification_factor|
-            stratifications = stratifications.product(stratification_factor.option_hashes)
-          end
-          stratifications.collect!(&:flatten)
-        end
-
-        stratifications.each do |option_hashes|
-          next if find_list_by_option_hashes(option_hashes)
-          stratification_factor_option_ids = option_hashes.collect { |oh| oh[:stratification_factor_option_id] }
-          options = @randomization_scheme.stratification_factor_options.where(id: stratification_factor_option_ids)
-          extra_options = option_hashes.select { |oh| oh[:extra] }
-          @randomization_scheme.lists.create(
-            project: @randomization_scheme.project,
-            user: current_user,
-            options: options,
-            extra_options: extra_options
-          )
-        end
-      end
-
-      def number_of_lists
-        @randomization_scheme.stratification_factors.collect { |sf| sf.option_hashes.count }.inject(:*).to_i
-      end
-
-      def find_list_by_criteria_pairs(criteria_pairs)
-        list = nil
-        @randomization_scheme.lists.each do |l|
-          if l.criteria_match?(criteria_pairs)
-            list = l
-            break
-          end
-        end
-        list
+        "Block size multipliers and treatment arms may not be set up correctly."
       end
 
       def randomize_subject_to_list!(subject, list, current_user, criteria_pairs)